AW: UMA metric for Safe Browsing manifest opt-in
No change to logic, this just adds UMA metrics.

This adds the "SafeBrowsing.WebView.AppOptIn" metric, which tracks
application opt-in preference for Safe Browsing. Most applications will
be in the "no opt-in prefernce" category, but it may be useful to see if
there's a change in "opt-out" or "opt-in" numbers, especially as we
begin to enable Safe Browsing by default.

This won't give us per-application metrics, but this should still
indicate (1) if a popular app(s) explicitly opts out of Safe Browsing,
and (2) a rough idea of how often users are protected by WebView Safe
Browsing (although this doesn't account for user opt-in preference).

Bug: 713724
Test: check that metric is logged under chrome://histograms/SafeBrowsing
Change-Id: I2fa01a7938eabd61aeacbb31f09e7729115254d3
Reviewed-on: https://chromium-review.googlesource.com/933290
Commit-Queue: Nate Fischer <ntfschr@chromium.org>
Reviewed-by: Ilya Sherman <isherman@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#538684}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4fd98c9e833e4af00a899ea68ae30e466477c6cf