desktop-pwas: Move WebContents with openers to new app windows.
Before, when we got a navigation for an in-scope URL we would cancel the
navigation, and start a new navigation in a new app window. This
resulted in the new app window losing its 'window.opener' and the
window that initiated the navigation losing its reference to the new
window.

Now, for initial navigations, we move the WebContents to a new app
window and continue the navigation as usual. This results in new
windows with `window.opener` keeping their openers.

Bug: 804881
Change-Id: Id0a32824cf7f9d57a9e32239e2e11ce952e91487
Reviewed-on: https://chromium-review.googlesource.com/923902
Commit-Queue: Giovanni Ortuño Urquidi <ortuno@chromium.org>
Reviewed-by: Matt Giuca <mgiuca@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#539079}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 69dc22e9a0ca25a5a32f62587e4226bcef21e94f