Added histogram for events of call API skew shifts in AEC3
The reporting to the histogram is added in the WebRTC CL
https://webrtc-review.googlesource.com/c/src/+/53066.

The metric for the histogram is the amount of skew shifts
detected for each minute in a call. It should be noted that
the metric is also reported for the case where there are 0
skew shifts. Furthermore, the metric is capped to 20 shifts
per minute.


Bug: chromium:814962
Change-Id: Ie6c7bb9b8e43c5b8ca8f77ac45e7b0eb2235aded
Reviewed-on: https://chromium-review.googlesource.com/933201
Reviewed-by: Mark Pearson <mpearson@chromium.org>
Reviewed-by: Ilya Sherman <isherman@chromium.org>
Commit-Queue: Per Åhgren <peah@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#539442}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f8b1134b1fded06591143ed43af3966ce9b4cff6