Add StateAtCancel UMA for downloads and offline pages
The histogram OfflineItems.StateAtCancel records the state of a request
at a user-initiated cancel. The name "OfflineItems" is used because the
metric is relevant to both downloads and offline pages. The histogram is
further broken down into two suffix histograms: Downloads and
OfflinePages.

The request state will be stored in the Intent associated with the
cancel button of notifications and Downloads Home. This ensures that the
cancel was user-initiated.

Bug: 796324
Change-Id: I9405ec90f62f00f8f532c973426ab094ee1e1e5d
Reviewed-on: https://chromium-review.googlesource.com/935969
Reviewed-by: Steven Holte <holte@chromium.org>
Reviewed-by: Cathy Li <chili@chromium.org>
Reviewed-by: Joy Ming <jming@chromium.org>
Commit-Queue: Candice Sy <cmsy@google.com>
Cr-Original-Commit-Position: refs/heads/master@{#539619}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0e75ff141f5cfe3e396c2847615e958b5162f37a