Show Message Center on lock screen behind flag.
This is a test bed for Lock Screen Notification
(go/notification-on-lock-screen).

This CL adds a new flag enable-lock-screen-notification. Behind this
flag, Message Center with notifications for the selected user is shown
on the lock screen.

Compared with Message Center on the non-locked screen, Message Center
on the lock screen does not have a settings button, which means it does
not have a separator between the settings button and a quiet mode button
either.

BUG=816353
TEST=ash_unittests

Change-Id: I27094a8a2c3c8b5f3b48f3ae5c8dccce4bfe78ab
Reviewed-on: https://chromium-review.googlesource.com/939223
Reviewed-by: Tetsui Ohkubo <tetsui@chromium.org>
Reviewed-by: James Cook <jamescook@chromium.org>
Commit-Queue: Megumi Hattori <megumihattori@google.com>
Cr-Original-Commit-Position: refs/heads/master@{#541067}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5a0d4181b538c577b034233bde90f4aa1a446aa2