cros: Remove "Tap dragging" accessibility settings
The setting already exists in Settings > Device > Touchpad and it's
confusing that it's in 2 places. It's also confusing that it causes
the accessibility person icon to appear in the status area.

* Remove the checkbox from a11y settings
* Update the a11y settings link to touchpad settings to mention
  tap dragging
* Take tap dragging out of system tray a11y submenu
* Don't show the a11y icon when tap dragging is enabled

Because webui settings are not available at the signin screen
the user will have to sign in before changing tap dragging. This
is consistent with other touchpad settings.

Bug: 164273, 724501
Test: ash_unittests and a11y browser_tests
Cq-Include-Trybots: master.tryserver.chromium.linux:closure_compilation
Change-Id: I65031173da7cab7e3f5dfc5efde98ec29fa625db
Reviewed-on: https://chromium-review.googlesource.com/971301
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
Reviewed-by: Ilya Sherman <isherman@chromium.org>
Reviewed-by: Qiang Xu <warx@google.com>
Commit-Queue: James Cook <jamescook@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#545200}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 827225bb7a05daa16d81e3d523cec30415281eb9