Added kiosk flag to flags histogram
Kiosk mode is available on Windows but there is no UMA to track it's
use, as the flag used to turn it on is missing from the histograms flags
enum.

The kiosk mode switch is called switches::kKioskMode in the code, and is
defined in chrome/common/chrome_switches.h.

Bug: 821572
Change-Id: Ice3a98e25c9d66103d9f9bbdd59c87d1587362a6
Reviewed-on: https://chromium-review.googlesource.com/972704
Commit-Queue: Ben Wells <benwells@chromium.org>
Reviewed-by: Matt Giuca <mgiuca@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#545350}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 60f581362d60dd3fd5e7f23e9217e5f85bcd38a3