Adding cross-origin metric for ad subframes.
This CL adds a metric that compares the number of ads that load in a
cross-origin subframe vs. the number that load in the same origin as the main
frame.  It counts any ad subframe or subframe of an ad in this metric.

Change-Id: I0b0a66c87ee981b3bbd54b86866b7f19195f6b3c
Bug: 970908
Reviewed-on: https://chromium-review.googlesource.com/970908
Reviewed-by: Alexei Svitkine <asvitkine@chromium.org>
Reviewed-by: Josh Karlin <jkarlin@chromium.org>
Reviewed-by: Charlie Harrison <csharrison@chromium.org>
Commit-Queue: Eric Robinson <ericrobinson@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#546194}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 81bbb440c7bd6f449b4f79b6a3db129fd98a14ae