Ensure metrics for preload=metadata suspend are accurate.
This adds three new metrics and fixes two existing ones to ensure
they are recorded accurately when we suspend for preload=metadata,
the new metrics are:

Media.PreloadMetadataSuspendWasIdeal.Audio
Media.PreloadMetadataSuspendWasIdeal.AudioVideo
Media.PreloadMetadataSuspendWasIdeal.Video

Which are boolean and set to true if a preload=metadata suspend is
not *immediately* aborted. Immediate abort can only result from
WMPI not knowing the true playback state until HAVE_FUTURE_DATA;
this is a lie we'd like to fix, and these metrics will help us
judge the importance of fixing it.

The fixed metrics are:
Media.HasEverPlayed
Media.TimeToFirstFrame.SRC
Media.TimeToPlayReady.SRC

HasEverPlayed is fixed by not counting the BUFFERING_HAVE_ENOUGH
event when it's issued for a suspended startup (determined by
a new media log property, which is gross, but how the existing
system must work for now).

TimeToFirstFrame and TimeToPlayReady are metrics which are relative
to the time of load start, but this does not work when we stop
after metadata. To ensure these metrics are still accurate once
playback does resume, we cache the |time_to_metadata_| value and
use it account for the section of time prior to metadata suspend.

BUG=694855, 827396
TEST=manually verified chrome://histograms has the right values.

Change-Id: I5cdeea08a798ed0875c2ce603d6ee2b9a31d1e55
Reviewed-on: https://chromium-review.googlesource.com/980589
Commit-Queue: Dale Curtis <dalecurtis@chromium.org>
Reviewed-by: Mark Pearson <mpearson@chromium.org>
Reviewed-by: Frank Liberato <liberato@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#547076}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ff57655e8674e9b167a062cec096b2c55a7da91d