[MediaRouter] Enterprise policy to allow all Cast IP addresses + cleanup
Added a new pref / policy for allowing (or disallowing) connecting to
Cast devices on public IPs, and check the pref when in
IsValidCastIPAddress().

Because the logic is now more complicated, consolidate the usage of the
function in a central place, i.e. CastMediaSinkServiceImpl. The
CastChannelAPI is no longer used as of M64 and should be removed soon.

CMSSImpl obtains PrefService from service manager and uses it to check
the "local state" (not associated with profile) pref.

Since getting the PrefService is async, it is technically possible to
try to open a channel before PrefService is available. To account
for this race condition case, we default to disallow all IPs if
PrefService is not yet available. The channel can be opened at a
later time when it becomes available (e.g., user gesture). This
should be a very rare corner case.

Also added TODOs to clean up unused code.

Bug: 813974
Change-Id: I0e1847463b1fb030f17f369d0d20bc09f81b2f7b
Reviewed-on: https://chromium-review.googlesource.com/982981
Commit-Queue: Derek Cheng <imcheng@chromium.org>
Reviewed-by: mark a. foltz <mfoltz@chromium.org>
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Bernhard Bauer <bauerb@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#548989}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 34478381b6adb0298c4103ea72e2c2b88aaa1502