Add RenderFrameHostImpl.DroppedInterfaceRequestName.
Interface requests to RenderFrame::GetRemoteInterfaces are dropped when
they arrive to the RenderFrameHostImpl after the RenderFrameHostImpl had
already committed the next cross-document navigation.

In follow-up to crrev.com/c/974263, this CL adds a histogram to record,
for each load, and for each dropped Mojo interface request in a frame, a
sample with a value corresponding to the hash of the dropped interface's
name, calculated as the lower 31 bits of the base::HashMetricName(name).

Bug: 795258
Change-Id: I0a4a3aa552c4bc2eb543b4aad3a3a35ecae9f831
Reviewed-on: https://chromium-review.googlesource.com/993059
Reviewed-by: Ilya Sherman <isherman@chromium.org>
Reviewed-by: Ken Rockot <rockot@chromium.org>
Reviewed-by: Nasko Oskov <nasko@chromium.org>
Commit-Queue: Balazs Engedy <engedy@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#549505}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: fa01f1ef035dc107d3796bda7102f06d52caea6b