Add tracker for signedin/syncing browsing activity
This is a resubmit with a fix of http://crrev/960005, after it broke
webgl tests. (see revert in http://crrev/1002854).

Modification since original CL:
In DesktopProfileSessionDurationService::OnSessionEnded, return early if
total_session_timer_ is null, which is the case when there was no
DesktopProfileSessionDurationService::OnSessionStarted call before it.

Original description:

This tracker reports the time that the user spent actively using chrome,
with their associated sync and sign-in state.

Definition of syncing: Having opted in to sync, and sync not being in an
"error" state because of a content-area sign-out.
Definition of signed in: The user is signed in to the content area.

These metrics will let us track:
 - How long on average users spend being signed in to the content area
 - How long users on average spend syncing
 - Which proportion of a user's time is being spent signed in
 - Which proportion of a user's time is being spent syncing

TBR reason: Small change to previously approved CL.

TBR=sdefresne@chromium.org, asvitkine@chromium.org

Bug: 820188
Change-Id: Ifbcd995d0d32ce4b967f9a6c215ba2fcc5477298
Reviewed-on: https://chromium-review.googlesource.com/1004675
Commit-Queue: Florian Uunk <feuunk@chromium.org>
Reviewed-by: David Roger <droger@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#549534}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 166aae480d77ac840b9c0b0ed04cfd7045f290cc