Record net::Error codes in LogDnsClient histograms
The old histogram, "DnsQueryStatus", summarised the net::Error codes as a handful of enum values.
This was unhelpful when a recent Finch experiment found that 75% of all queries were failing for
reason "unknown". By recording the actual net::Error code, hopefully a little more light will be
cast on the actual reason for the failures. Two histograms will be used - one for failures while
obtaining the leaf index, and the other for the result of obtaining the audit proof.

Bug: 506227
Change-Id: If143056dd284598b31152d57a41d5a86a9c53408
Reviewed-on: https://chromium-review.googlesource.com/1010483
Commit-Queue: Rob Percival <robpercival@chromium.org>
Reviewed-by: Ilya Sherman <isherman@chromium.org>
Reviewed-by: Ryan Sleevi <rsleevi@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#550537}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 7976c213670c83545b92161c5c33c3e64fbfa4fb