Adding a MeasuredURLLoader which will wrap SimpleURLLoader
This class will be consuemd by all Chrome services/features for the purposes of
tracking data usage and other statistics about loading behavior.

This CL introduces the class, but does not add any consumers, and the
functionality to track loads is not yet added.

The purpose of MeasuredURLLoader is to provide a loading API for Chrome
services that will track metrics around their data use to monitor if any
services have excessive data use. The expectation is that most
instances of URLFetcher will be transitioned to ServiceURLLoader with
any loads that should be untracked being moved to SimpleURLLoader.

Bug: 808498
Change-Id: I6fc75e37627272e23e96c3341d3593a185c0bf36
Reviewed-on: https://chromium-review.googlesource.com/899822
Commit-Queue: Ryan Sturm <ryansturm@chromium.org>
Reviewed-by: Robert Kaplow <rkaplow@chromium.org>
Reviewed-by: Tarun Bansal <tbansal@chromium.org>
Reviewed-by: Matt Menke <mmenke@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#551905}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 1de0a8de73b4c1acea7ee5dd72fbbd1b782e6020