Fix two bugs in the Mac implementation of sizes.py.
1) If Xcode's command line tools were not installed, otool would fail to
run. However the script was not checking the return code of the command
and the script would silently fail.

2) The static initializer count assumed a 32-bit word size. Mac has been
exclusively 64-bit for years, so assume an 8-byte word size.

This also re-lowers the expected count of static initializers now that
the ones that crept in have been removed. See the bug for details.

Bug: 831951
Change-Id: If93b5603531d6a0933a04d67b55b7559869c1266
Reviewed-on: https://chromium-review.googlesource.com/1018440
Reviewed-by: Dirk Pranke <dpranke@chromium.org>
Commit-Queue: Robert Sesek <rsesek@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#551997}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 29c584bb596312da4300125833885170b9b58612