Add a new webstore_private API to get the referrer chain.
Like for inline-installation, the Chrome Web Store should be able to get
the full referrer chain of how the user got to the CWS item detail page.
This change adds a new API that the CWS (and only the CWS) will be able
to call in order to report & detect misleading user acquisition patterns.

Note that the SafeBrowsing "download" ping which is sent if the user
chooses to install the extension already contains the referrer data. This
change effectively allows the CWS to collect the data before the user
chooses to install the extension.

Bug: 807090
Change-Id: I6b7f85fca75fc134a3d6207cb47a1f3674d15c96
Reviewed-on: https://chromium-review.googlesource.com/1015792
Commit-Queue: proberge <proberge@chromium.org>
Reviewed-by: Devlin <rdevlin.cronin@chromium.org>
Reviewed-by: Jialiu Lin <jialiul@chromium.org>
Reviewed-by: Gayane Petrosyan <gayane@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#552441}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: bab8de29fb63e3fa92b3fd888a1242bfde844f45