Clean up TerminationStatus UMA enum by deprecating histogram
base::TerminationStatus is used as an enum in the
GPU.GPUProcessTerminationStatus histogram. This is not ideal because
* it forces base::TerminationStatus values to be stable across versions
* base::TerminationStatus has OS-defines that make values differ
  depending on the platform, which makes the enum description
  complex.

Create a new GpuTerminationStatus that does not have OS defines.
Then depcreate GPUProcessTerminationStatus and create a new histogram
that uses the new enum instead.

Change-Id: I768230b9e690004c43f0487a5d00f11d8806bf61
Reviewed-on: https://chromium-review.googlesource.com/1011549
Commit-Queue: Bo <boliu@chromium.org>
Reviewed-by: Steven Holte <holte@chromium.org>
Reviewed-by: Kenneth Russell <kbr@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#552645}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 82d3cc6b3ab2f2140e1b32e7b4bbac3bd20f2b3d