[VK] Add private OnFocus event with gesture information.
This patch adds a new OnFocus event to the private keyboard API. This
event has additional information about whether the focus was triggered
by a pen stylus or other means.

The plan is for IME extension to listen to this event and change its
keyset/position/mode etc. depending on the focus reason.

If the extension listens to the private OnFocus event, then the public
version would not be triggered (and vice versa). This is to prevent
OnFocus from being raised twice.

Bug: 824604
Change-Id: I5dd523a183d517e166bbc73abd6a79a6557c6942
Reviewed-on: https://chromium-review.googlesource.com/1015446
Commit-Queue: Darren Shen <shend@chromium.org>
Reviewed-by: Sadrul Chowdhury <sadrul@chromium.org>
Reviewed-by: Michael Wasserman <msw@chromium.org>
Reviewed-by: Yusuke Sato <yusukes@chromium.org>
Reviewed-by: Steven Holte <holte@chromium.org>
Reviewed-by: Shu Chen <shuchen@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#555997}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: dd52d4ae0676b31b56155aa91e5fa2d278c45f21