Remove extra patterns from Traffic Annotation Extractor.
Socket patterns are no longer required in traffic annotation tests,
hence they are removed from traffic_annotation_extractor clang tool.

The original plan was to have network traffic annotation as an optional
argument of socket create functions, and we needed to check if the
argument existed for all calls that originate from Chrome. This is now
changed to having annotations as a mandatory argument of the socket
write functions, and as it is enforced by the compiler, we do not need
to test it.

Bug: 690323
Change-Id: I101fb7295662c3488807091cb9a8686cfc87d249
Reviewed-on: https://chromium-review.googlesource.com/1046087
Commit-Queue: Ramin Halavati <rhalavati@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#556388}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f16d9ea8fc7f52b8ace64c1dafe6d77bcc0d155d